WIP: move components into ui lib
