+++ projects/app/src/app/__demo__/pages/autocomplete/autocomplete.component.ts
@@ -7,9 +7,9 @@ import {
   Validators,
 } from '@angular/forms';
 import { Observable, delay, of } from 'rxjs';
+import { TextInputComponent } from '@ui/components/text-input';
 
 import {
-  TextInputComponent,
   FormFieldComponent,
   AutocompleteAsyncOptionsFn,
   AutocompleteOption,
