+++ projects/app/src/app/app.component.ts
@@ -4,9 +4,9 @@ import { TranslocoModule } from '@jsverse/transloco';
 import { LinearSpinnerComponent } from '@ui/components/linear-spinner';
 import { ModalHostComponent } from '@ui/components/modal';
 
-import { NotificationsHostComponent } from '@app/common/components';
+import { NotificationsHostComponent } from '@ui/components/notification';
 import { SoftwareUpdateService } from '@app/core/sw-update';
-import { UiStore } from '@app/core/ui';
+import { UiStore, NOTIFICATION_TIMEOUT } from '@app/core/ui';
 import { UserStore } from '@app/features/user/store/feature';
 
 @Component({
@@ -26,6 +26,7 @@ export class AppComponent implements OnInit {
   private swUpdate = inject(SoftwareUpdateService);
   private userStore = inject(UserStore);
 
+  NOTIFICATION_TIMEOUT = NOTIFICATION_TIMEOUT;
   notification = this.uiStore.notifications.notification;
   loading = this.uiStore.loader.loading;
   themeConfig = this.uiStore.theme.config;
