+++ projects/app/src/app/common/components/index.ts
@@ -1,8 +1,5 @@
 export * from './autocomplete';
 export * from './change-category-modal';
-export * from './confirm-prompt-modal';
 export * from './form-field';
-export * from './notification';
-export * from './text-input';
 export * from './textarea';
 export * from './user-credentials-form';
