+++ projects/app/src/app/features/list/components/item-form-modal/item-form-modal.component.ts
@@ -27,13 +27,13 @@ import {
   ModalFooterDirective,
   ModalHeaderDirective,
 } from '@ui/components/modal';
+import { TextInputComponent } from '@ui/components/text-input';
 
 import {
   AUTOCOMPLETE_EXPORTS,
   AutocompleteAsyncOptionsFn,
   AutocompleteOption,
   FORM_FIELD_EXPORTS,
-  TextInputComponent,
   TextareaComponent,
 } from '@app/common/components';
 import { DEFAULT_CATEGORY } from '@app/core/constants';
