+++ projects/app/src/app/features/user/pages/signup/signup.component.ts
@@ -12,10 +12,11 @@ import { TranslocoModule } from '@jsverse/transloco';
 import { finalize } from 'rxjs';
 import { ButtonComponent } from '@ui/components/button';
 import { PageHeaderComponent } from '@ui/components/page-header';
+import { TextInputComponent } from '@ui/components/text-input';
 
 import { DEFAULT_ROUTE } from '@app/app.routes';
 import { UiStore } from '@app/core/ui';
-import { FORM_FIELD_EXPORTS, TextInputComponent } from '@app/common/components';
+import { FORM_FIELD_EXPORTS } from '@app/common/components';
 import { FIELD_PIPES_EXPORTS } from '@ui/pipes';
 import { getFieldDescriptor as fDescribe } from '@common/utils';
 import { InvitesService } from '../../services';
