+++ projects/ui/src/components/confirm-prompt-modal/confirm-prompt-modal.component.ts
