+++ projects/ui/src/components/confirm-prompt-modal/public-api.ts
