+++ projects/ui/src/components/notification/notifications-host/notifications-host.component.html
@@ -2,7 +2,7 @@
 <app-notification
   [notificationId]="notification!.id"
   [notificationType]="notification!.type"
-  [dismissAfter]="NOTIFICATION_TIMEOUT"
+  [dismissAfter]="dismissAfter()"
   [more]="notification!.more"
   (dismissed)="dismissed.emit()"
 >
