+++ projects/ui/src/components/public-api.ts
@@ -3,10 +3,13 @@ export * from './bottom-menu/public-api';
 export * from './button/public-api';
 export * from './card-list/public-api';
 export * from './checkbox/public-api';
+export * from './confirm-prompt-modal/public-api';
 export * from './icon-button/public-api';
 export * from './linear-spinner/public-api';
 export * from './modal/public-api';
+export * from './notification/public-api';
 export * from './page-header/public-api';
 export * from './quick-number/public-api';
 export * from './select/public-api';
+export * from './text-input/public-api';
 export * from './toggle/public-api';
