+++ projects/ui/src/components/text-input/public-api.ts
